requirements.txt
pgn2data
chess
python-lichess